Fix pgdg package name for 8.4
The package names for PostgreSQL 8.4 follow a slightly different naming
pattern. Fixed the matching logic in get_repo_rpm_release.py.

Also changed the script to output the whole package name, which is then
used in the install task. This is to prevent having to fiddle with jinja
filters in the install task to get the correct download url.