Add RHEL GPG key with PGDG prefix (allows installing PostgreSQL >= 11)
According to https://download.postgresql.org/pub/repos/yum/keys/README.txt, there are two types of keys for RHEL: `RPM-GPG-KEY-PGDG*` and `PGDG-RPM-GPG-KEY*`.

> This directory contains keys used for signing the PostgreSQL RPMs on RHEL (and its derivatives) and Fedora.
> The keys which have the PGDG prefix are used to sign PostgreSQL 11 and above RPMs.
>
> Archived versions (v10 and below) use the RPM-GPG-KEY prefix.

The role is already adding the first kind of key to the system, but not the second. Therefore, installing newer PostgreSQL versions is not possible at the moment. This commit adds the second type of key so that newer PostgreSQL versions can be installed too.